requirements.txt
Flask
google-generativeai
SpeechRecognition
